requirements.txt
alembic==1.7.5
amqp==5.0.6
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.2
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.4.60
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.28
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
